requirements.txt
Flask==3.0.0
Flask-SocketIO==5.3.6
cryptography==41.0.7
stegano>=1.0.0
Pillow
eventlet==0.33.3
pyngrok==7.1.0
